requirements.txt
httpx>=0.27.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
openai>=1.0.0
pyyaml>=6.0
